requirements.txt
# for Heroku
-r requirements/development.txt
django-toolbelt